added a few prototypes

+++ osa5/bloglist-frontend/src/App.js
@@ -102,9 +102,11 @@ const App = () => {
 
 	const handleRemove = async id => {
 		const blog = blogs.find(b => b.id === id)
-		if (window.confirm(`are you sure you want to delete ${blog.title} by ${blog.author}`)) {
-			await blogService.remove(blog)
-			setBlogs(blogs.filter(blog => blog.id !== id))
+		if (user.username === blog.user.username) {
+			if (window.confirm(`are you sure you want to delete ${blog.title} by ${blog.author}`)) {
+				await blogService.remove(blog)
+				setBlogs(blogs.filter(blog => blog.id !== id))
+			}
 		}
 	}
